Full CRUD functionality had been created

+++ mysite/blog/views.py
@@ -74,42 +74,14 @@ def logout_request(request):
     return redirect("/home/")
 
 
-# @login_required
-# def delete_comment(request, comment_id):
-#     if request.method == 'POST':
-#         comment = Comment.objects.get(id=comment_id)
-#         if comment.name == request.user.username:
-#             comment.delete()
-#             return redirect(request.META.get('HTTP_REFERER'))
-#         else:
-#             return HttpResponse("You are not authorized to delete this comment", status=401)
-#     else:
-#         return HttpResponse("Method not allowed", status=405)
-
-
-# @login_required
-# def delete_comment(request):
-#     if request.method == 'POST':
-#         comment_id = request.POST.get('comment_id')
-#         comment = Comment.objects.get(id=comment_id)
-#         if comment.name == request.user.username:
-#             comment.delete()
-#             return redirect(request.META.get('HTTP_REFERER'))
-#         else:
-#             return HttpResponse("You are not authorized to delete this comment", status=401)
-
-
-
 def delete_comment(request):
-    comments = Comment.objects.filter(name=request.user.username)
-    print(comments)
-    if request.method == 'POST':
-        # comment_name = request.POST.get('Kate')
-        comment = Comment.objects.get(name="AD")
-        comment.delete()
-        return redirect(request.META.get('HTTP_REFERER'))
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            comment = Comment.objects.filter(name=request.user.username).first()
+            comment.delete()
+            return redirect(request.META.get('HTTP_REFERER'))
     else:
-        return HttpResponse("Method not allowed", status=405)
+        return HttpResponse("You need to be logged in to delete a comment", status=405)
 
 
 def register(request):
